refactor(#3098): extract ensureNativeArrayHof into src/codegen/hof-native.ts — satisfy the #3102 LOC-regrowth ratchet ✓
The new LOC-regrowth gate (landed mid-flight in #2812's wave) correctly
rejected +368 lines in object-runtime.ts. Move the HOF-loop emitter into its
own subsystem module per the consolidation plan (object-runtime owns the MOP
substrate; hof-native owns the callback-consuming loops built on it).
object-runtime.ts is byte-identical to origin/main again. calls.ts +22
(the call-site gate itself) banked via --update, visible here in review.
coercion-sites baseline regenerated from the merged tree: hof-native.ts +1
(the __is_truthy consumer call, moved out of object-runtime), and
property-access.ts 16→17 — the same correction PR #2814 carries for the
post-#2812 stale-checkout clobber; both sides agree so the merge is trivial.

Behavior identical: repro probes + all 15 issue-3098 tests re-verified after
the extraction.
