+++ scripts/coercion-sites-baseline.json
@@ -18,6 +18,7 @@
   "codegen/expressions.ts": 1,
   "codegen/fixups.ts": 1,
   "codegen/function-body.ts": 1,
+  "codegen/hof-native.ts": 1,
   "codegen/host-import-allowlist.ts": 2,
   "codegen/index.ts": 35,
   "codegen/iterator-native.ts": 2,
@@ -25,7 +26,7 @@
   "codegen/map-runtime.ts": 2,
   "codegen/number-format-native.ts": 8,
   "codegen/object-ops.ts": 4,
-  "codegen/object-runtime.ts": 28,
+  "codegen/object-runtime.ts": 27,
   "codegen/parse-number-native.ts": 4,
   "codegen/promise-combinators.ts": 1,
   "codegen/property-access.ts": 17,
